Introducing the chaos_function that causes servers to randomly crash.
New solution needs to be implemented for obtaining a quarom though.  The
loop causes the quarom to take too long for the timeout to expire which
causes the client to stop waiting.  (timeouts are equal on both sides
and we have to go one level lower into RPC do to this.

Might be something we just want to put into the executive sommary.